Move tests into their own directory, autogenerate fingerprint tests.
