+++ .gitignore
@@ -6,3 +6,6 @@ postgres.tar.bz2
 
 examples/*
 !examples/*.c
+
+test/*
+!test/*.c
